requirements.txt
python-dotenv
pymupdf
python-docx
sqlalchemy
psycopg[binary]
pgvector
google-genai
